# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EventFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventInputStatus = typeof EventInputStatus[keyof typeof EventInputStatus];


export const EventInputStatus = {
  planning: 'planning',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventInput {
  name: string;
  category: string;
  eventType: string;
  date: string;
  time?: string;
  city: string;
  state: string;
  venue: string;
  organizer: string;
  capacity: number;
  status: EventInputStatus;
}

export type Event = EventInput & {
  id: string;
  createdAt: string;
};

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export type TransactionInputStatus = typeof TransactionInputStatus[keyof typeof TransactionInputStatus];


export const TransactionInputStatus = {
  pending: 'pending',
  paid: 'paid',
  overdue: 'overdue',
  cancelled: 'cancelled',
} as const;

export type TransactionInputPaymentMethod = typeof TransactionInputPaymentMethod[keyof typeof TransactionInputPaymentMethod];


export const TransactionInputPaymentMethod = {
  pix: 'pix',
  credit_card: 'credit_card',
  boleto: 'boleto',
  bank_transfer: 'bank_transfer',
  cash: 'cash',
} as const;

export interface TransactionInput {
  eventId: string;
  type: TransactionInputType;
  category: string;
  description: string;
  amount: number;
  dueDate: string;
  status: TransactionInputStatus;
  paymentMethod: TransactionInputPaymentMethod;
}

export type Transaction = TransactionInput & {
  id: string;
  createdAt: string;
};

export interface DashboardSummary {
  activeEventsCount: number;
  totalRegistrations: number;
  expectedRevenue: number;
  realizedRevenue: number;
  totalExpenses: number;
  profit: number;
  sponsorsCount: number;
  suppliersCount: number;
  staffCount: number;
  pendingContracts: number;
  contractsExpiringSoon: number;
  upcomingEvents: Event[];
  recentTransactions: Transaction[];
}

export type SponsorInputTier = typeof SponsorInputTier[keyof typeof SponsorInputTier];


export const SponsorInputTier = {
  master: 'master',
  gold: 'gold',
  silver: 'silver',
  bronze: 'bronze',
  naming_rights: 'naming_rights',
} as const;

export type SponsorInputStatus = typeof SponsorInputStatus[keyof typeof SponsorInputStatus];


export const SponsorInputStatus = {
  prospect: 'prospect',
  negotiating: 'negotiating',
  confirmed: 'confirmed',
  delivered: 'delivered',
} as const;

export interface SponsorInput {
  eventId: string;
  name: string;
  tier: SponsorInputTier;
  contactName: string;
  email: string;
  phone: string;
  investment: number;
  status: SponsorInputStatus;
}

export type Sponsor = SponsorInput & {
  id: string;
  createdAt: string;
};

export type SupplierInputStatus = typeof SupplierInputStatus[keyof typeof SupplierInputStatus];


export const SupplierInputStatus = {
  active: 'active',
  inactive: 'inactive',
  pending_review: 'pending_review',
} as const;

export interface SupplierInput {
  name: string;
  category: string;
  contactName: string;
  email: string;
  phone: string;
  price: number;
  rating: number;
  status: SupplierInputStatus;
}

export type Supplier = SupplierInput & {
  id: string;
  createdAt: string;
};

export type ContractInputContractType = typeof ContractInputContractType[keyof typeof ContractInputContractType];


export const ContractInputContractType = {
  supplier: 'supplier',
  sponsor: 'sponsor',
  staff: 'staff',
  artist: 'artist',
  speaker: 'speaker',
  exhibitor: 'exhibitor',
  volunteer: 'volunteer',
} as const;

export type ContractInputStatus = typeof ContractInputStatus[keyof typeof ContractInputStatus];


export const ContractInputStatus = {
  draft: 'draft',
  sent: 'sent',
  signed: 'signed',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export interface ContractInput {
  eventId: string;
  title: string;
  contractType: ContractInputContractType;
  counterpartyName: string;
  value: number;
  startDate: string;
  endDate: string;
  status: ContractInputStatus;
}

export type Contract = ContractInput & {
  id: string;
  createdAt: string;
};

export type StaffMemberInputStatus = typeof StaffMemberInputStatus[keyof typeof StaffMemberInputStatus];


export const StaffMemberInputStatus = {
  scheduled: 'scheduled',
  checked_in: 'checked_in',
  checked_out: 'checked_out',
  no_show: 'no_show',
} as const;

export interface StaffMemberInput {
  eventId: string;
  name: string;
  role: string;
  email: string;
  phone: string;
  hourlyRate: number;
  status: StaffMemberInputStatus;
}

export type StaffMember = StaffMemberInput & {
  id: string;
  createdAt: string;
};

export type RegistrationInputTicketType = typeof RegistrationInputTicketType[keyof typeof RegistrationInputTicketType];


export const RegistrationInputTicketType = {
  free: 'free',
  paid: 'paid',
  vip: 'vip',
  box: 'box',
  premium: 'premium',
  combo: 'combo',
} as const;

export type RegistrationInputStatus = typeof RegistrationInputStatus[keyof typeof RegistrationInputStatus];


export const RegistrationInputStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  cancelled: 'cancelled',
  waitlist: 'waitlist',
} as const;

export interface RegistrationInput {
  eventId: string;
  participantName: string;
  email: string;
  phone: string;
  category: string;
  ticketType: RegistrationInputTicketType;
  price: number;
  status: RegistrationInputStatus;
  checkedIn: boolean;
}

export type Registration = RegistrationInput & {
  id: string;
  createdAt: string;
};

export type ListSponsorsParams = {
eventId?: string;
};

export type ListSuppliersParams = {
category?: string;
};

export type ListContractsParams = {
eventId?: string;
};

export type ListStaffParams = {
eventId?: string;
};

export type ListTransactionsParams = {
eventId?: string;
type?: ListTransactionsType;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
} as const;

export type ListRegistrationsParams = {
eventId?: string;
};