# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      responses:
        "204":
          description: Deleted
  /sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors]
      summary: List sponsors
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of sponsors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
    post:
      operationId: createSponsor
      tags: [sponsors]
      summary: Create a sponsor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "201":
          description: Created sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getSponsor
      tags: [sponsors]
      summary: Get a sponsor
      responses:
        "200":
          description: Sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    put:
      operationId: updateSponsor
      tags: [sponsors]
      summary: Update a sponsor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "200":
          description: Updated sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    delete:
      operationId: deleteSponsor
      tags: [sponsors]
      summary: Delete a sponsor
      responses:
        "204":
          description: Deleted
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    put:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      responses:
        "204":
          description: Deleted
...
  /registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List registrations
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Create a registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Created registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get a registration
      responses:
        "200":
          description: Registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
    put:
      operationId: updateRegistration
      tags: [registrations]
      summary: Update a registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "200":
          description: Updated registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
    delete:
      operationId: deleteRegistration
      tags: [registrations]
      summary: Delete a registration
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    EventInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        eventType:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        city:
          type: string
        state:
          type: string
        venue:
          type: string
        organizer:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [planning, confirmed, in_progress, completed, cancelled]
      required:
        - name
        - category
        - eventType
        - date
        - city
        - state
        - venue
        - organizer
        - capacity
        - status
    Event:
      allOf:
        - $ref: "#/components/schemas/EventInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
    SponsorInput:
      type: object
      properties:
        eventId:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [master, gold, silver, bronze, naming_rights]
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        investment:
          type: number
        status:
          type: string
          enum: [prospect, negotiating, confirmed, delivered]
      required:
        - eventId
        - name
        - tier
        - contactName
        - email
        - phone
        - investment
        - status
    Sponsor:
      allOf:
        - $ref: "#/components/schemas/SponsorInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
...
    ContractInput:
      type: object
      properties:
        eventId:
          type: string
        title:
          type: string
        contractType:
          type: string
          enum: [supplier, sponsor, staff, artist, speaker, exhibitor, volunteer]
        counterpartyName:
          type: string
        value:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, sent, signed, expired, cancelled]
      required:
        - eventId
        - title
        - contractType
        - counterpartyName
        - value
        - startDate
        - endDate
        - status
    Contract:
      allOf:
        - $ref: "#/components/schemas/ContractInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
...
    RegistrationInput:
      type: object
      properties:
        eventId:
          type: string
        participantName:
          type: string
        email:
          type: string
        phone:
          type: string
        category:
          type: string
        ticketType:
          type: string
          enum: [free, paid, vip, box, premium, combo]
        price:
          type: number
        status:
          type: string
          enum: [confirmed, pending, cancelled, waitlist]
        checkedIn:
          type: boolean
      required:
        - eventId
        - participantName
        - email
        - phone
        - category
        - ticketType
        - price
        - status
        - checkedIn
    Registration:
      allOf:
        - $ref: "#/components/schemas/RegistrationInput"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt