# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EventFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInput } from './registrationInput';

export type Registration = RegistrationInput & {
  id: string;
  createdAt: Date;
};