# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EventFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputStatus } from './registrationInputStatus';
import type { RegistrationInputTicketType } from './registrationInputTicketType';

export interface RegistrationInput {
  eventId: string;
  participantName: string;
  email: string;
  phone: string;
  category: string;
  ticketType: RegistrationInputTicketType;
  price: number;
  status: RegistrationInputStatus;
  checkedIn: boolean;
}