# lib/api-zod/src/generated/types/sponsor.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EventFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorInput } from './sponsorInput';

export type Sponsor = SponsorInput & {
  id: string;
  createdAt: Date;
};